+++ src/shared/ui/calendar/index.ts
@@ -1,1 +1,2 @@
 export { Calendar } from "./calendar";
+export { CalendarDayButton } from "./calendar-day-button";
